misc double output (#797)
* convert decimal to double for MACD, PVO, and CHOP

+++ src/m-r/Macd/Macd.Models.cs
@@ -3,9 +3,9 @@ namespace Skender.Stock.Indicators;
 [Serializable]
 public class MacdResult : ResultBase
 {
-    public decimal? Macd { get; set; }
-    public decimal? Signal { get; set; }
-    public decimal? Histogram { get; set; }
+    public double? Macd { get; set; }
+    public double? Signal { get; set; }
+    public double? Histogram { get; set; }
 
     // extra interim data
     public decimal? FastEma { get; set; }
